Separate trace API components into own files
This is moves the Span (and related types), Tracer, and TracerProvider
out of the single trace.go and into their own files. This change is
intended to help developers find these types when looking at the
directory and to split apart the long trace.go file.
